Unify interop-testing with io.grpc and akka-scala server
To more easily be able to add akka-java server later
