+++ interop-tests/src/main/scala/com/lightbend/grpc/interop/AkkaGrpcServerScala.scala
@@ -1,0 +1,77 @@
+package com.lightbend.grpc.interop
+
+import java.io.FileInputStream
+import java.nio.file.{ Files, Paths }
+import java.security.cert.CertificateFactory
+import java.security.spec.PKCS8EncodedKeySpec
+import java.security.{ KeyFactory, KeyStore, SecureRandom }
+import java.util.Base64
+import javax.net.ssl.{ KeyManagerFactory, SSLContext }
+
+import akka.actor.ActorSystem
+import akka.http.scaladsl.Http.ServerBinding
+import akka.http.scaladsl.model.{ HttpRequest, HttpResponse, StatusCodes }
+import akka.http.scaladsl.{ Http2, HttpsConnectionContext }
+import akka.stream.{ ActorMaterializer, Materializer }
+import io.grpc.internal.testing.TestUtils
+
+import scala.concurrent.duration._
+import scala.concurrent.{ Await, ExecutionContext, Future }
+
+case class AkkaGrpcServerScala(serverHandlerFactory: Materializer => ExecutionContext => PartialFunction[HttpRequest, Future[HttpResponse]]) extends GrpcServer[(ActorSystem, ServerBinding)] {
+  override def start() = {
+    implicit val sys = ActorSystem()
+    implicit val mat = ActorMaterializer()
+    implicit val ec = sys.dispatcher
+
+    val testService = serverHandlerFactory(mat)(ec)
+
+    val bindingFuture = Http2().bindAndHandleAsync(
+      testService.orElse { case _: HttpRequest ⇒ Future.successful(HttpResponse(StatusCodes.NotFound)) },
+      interface = "127.0.0.1",
+      port = 8080,
+      httpsContext = serverHttpContext())
+
+    val binding = Await.result(bindingFuture, 10.seconds)
+    (sys, binding)
+  }
+
+  override def stop(binding: (ActorSystem, ServerBinding)) = binding match {
+    case (sys, binding) ⇒
+      sys.log.info("Exception thrown, unbinding")
+      Await.result(binding.unbind(), 10.seconds)
+      Await.result(sys.terminate(), 10.seconds)
+
+  }
+
+  private def serverHttpContext() = {
+    val keyEncoded = new String(Files.readAllBytes(Paths.get(TestUtils.loadCert("server1.key").getAbsolutePath)), "UTF-8")
+      .replace("-----BEGIN PRIVATE KEY-----\n", "")
+      .replace("-----END PRIVATE KEY-----\n", "")
+      .replace("\n", "")
+
+    val decodedKey = Base64.getDecoder.decode(keyEncoded)
+
+    val spec = new PKCS8EncodedKeySpec(decodedKey)
+
+    val kf = KeyFactory.getInstance("RSA")
+    val privateKey = kf.generatePrivate(spec)
+
+    val fact = CertificateFactory.getInstance("X.509")
+    val is = new FileInputStream(TestUtils.loadCert("server1.pem"))
+    val cer = fact.generateCertificate(is)
+
+    val ks = KeyStore.getInstance("PKCS12")
+    ks.load(null)
+    ks.setKeyEntry("private", privateKey, Array.empty, Array(cer))
+
+    val keyManagerFactory = KeyManagerFactory.getInstance("SunX509")
+    keyManagerFactory.init(ks, null)
+
+    val context = SSLContext.getInstance("TLS")
+    context.init(keyManagerFactory.getKeyManagers, null, new SecureRandom)
+
+    new HttpsConnectionContext(context)
+  }
+
+}
