+++ interop-tests/src/main/scala/com/lightbend/grpc/interop/GrpcInteropTests.scala
@@ -1,25 +1,11 @@
 package com.lightbend.grpc.interop
 
-import java.io.FileInputStream
-import java.nio.file.{Files, Paths}
-import java.security.{KeyFactory, KeyStore, SecureRandom}
-import java.security.cert.CertificateFactory
-import java.security.spec.PKCS8EncodedKeySpec
-import java.util.Base64
-import javax.net.ssl.{KeyManagerFactory, SSLContext}
-
-import akka.actor.ActorSystem
-import akka.http.scaladsl.{Http2, HttpsConnectionContext}
-import akka.http.scaladsl.model.{HttpRequest, HttpResponse, StatusCodes}
-import akka.stream.{ActorMaterializer, Materializer}
 import io.grpc.StatusRuntimeException
-import io.grpc.internal.testing.TestUtils
 import io.grpc.testing.integration.Util
 import io.grpc.testing.integration2._
 import org.scalatest.{Assertion, Succeeded, WordSpec}
 
-import scala.concurrent.duration._
-import scala.concurrent.{Await, ExecutionContext, Future}
+import scala.concurrent.ExecutionContext
 import scala.util.control.NonFatal
 
 trait GrpcInteropTests { self: WordSpec =>
@@ -46,28 +32,13 @@ trait GrpcInteropTests { self: WordSpec =>
     "unimplemented_service",
   )
 
-
-  def javaGrpcTests(clientTesterProvider: ClientTesterProvider) =
-    GrpcJavaServer.label + " with " + clientTesterProvider.label should {
-      testCases.foreach { testCaseName =>
-        s"pass the $testCaseName integration test" in {
-          val allPending = GrpcJavaServer.pendingCases ++ clientTesterProvider.pendingCases
-          pendingTestCaseSupport(allPending(testCaseName)) {
-            withGrpcJavaServer() {
-              runGrpcClient(testCaseName)(clientTesterProvider.clientTesterFactory)
-            }
-          }
-        }
-      }
-    }
-
-  def akkaGrpcTests(serverHandlerProvider: ServerHandlerProvider, clientTesterProvider: ClientTesterProvider) =
+  def grpcTests(serverHandlerProvider: ServerHandlerProvider, clientTesterProvider: ClientTesterProvider) =
     serverHandlerProvider.label + " with " + clientTesterProvider.label should {
       testCases.foreach { testCaseName =>
         s"pass the $testCaseName integration test" in {
           val allPending = serverHandlerProvider.pendingCases ++ clientTesterProvider.pendingCases
           pendingTestCaseSupport(allPending(testCaseName)) {
-            withGrpcAkkaServer(serverHandlerProvider.serverHandlerFactory) {
+            withGrpcServer(serverHandlerProvider.server) {
               runGrpcClient(testCaseName)(clientTesterProvider.clientTesterFactory)
             }
           }
@@ -105,27 +76,13 @@ trait GrpcInteropTests { self: WordSpec =>
     }
   }
 
-  private def withGrpcAkkaServer(testServiceFactory: Materializer => ExecutionContext => PartialFunction[HttpRequest, Future[HttpResponse]])(block: => Unit): Assertion = {
-    implicit val sys = ActorSystem()
+  private def withGrpcServer[T](server: GrpcServer[T])(block: => Unit): Assertion = {
     try {
-      implicit val mat = ActorMaterializer()
-
-      val testService = testServiceFactory(mat)(sys.dispatcher)
-
-      val bindingFuture = Http2().bindAndHandleAsync(
-        testService.orElse { case _: HttpRequest ⇒ Future.successful(HttpResponse(StatusCodes.NotFound)) },
-        interface = "127.0.0.1",
-        port = 8080,
-        httpsContext = serverHttpContext())
-
-      val binding = Await.result(bindingFuture, 10.seconds)
-
+      val binding = server.start()
       try {
         block
       } finally {
-        sys.log.info("Exception thrown, unbinding")
-        Await.result(binding.unbind(), 10.seconds)
-        Await.result(sys.terminate(), 10.seconds)
+        server.stop(binding)
       }
       Succeeded
     } catch {
@@ -134,55 +91,6 @@ trait GrpcInteropTests { self: WordSpec =>
         // to avoid trouble when running tests from sbt
         if (e.getCause == null) fail(e.getMessage)
         else fail(e.getMessage, e.getCause)
-      case NonFatal(t) => fail(t)
-    }
-  }
-
-  private def serverHttpContext() = {
-    val keyEncoded = new String(Files.readAllBytes(Paths.get(TestUtils.loadCert("server1.key").getAbsolutePath)), "UTF-8")
-      .replace("-----BEGIN PRIVATE KEY-----\n", "")
-      .replace("-----END PRIVATE KEY-----\n", "")
-      .replace("\n", "")
-
-    val decodedKey = Base64.getDecoder.decode(keyEncoded)
-
-    val spec = new PKCS8EncodedKeySpec(decodedKey)
-
-    val kf = KeyFactory.getInstance("RSA")
-    val privateKey = kf.generatePrivate(spec)
-
-    val fact = CertificateFactory.getInstance("X.509")
-    val is = new FileInputStream(TestUtils.loadCert("server1.pem"))
-    val cer = fact.generateCertificate(is)
-
-    val ks = KeyStore.getInstance("PKCS12")
-    ks.load(null)
-    ks.setKeyEntry("private", privateKey, Array.empty, Array(cer))
-
-    val keyManagerFactory = KeyManagerFactory.getInstance("SunX509")
-    keyManagerFactory.init(ks, null)
-
-    val context = SSLContext.getInstance("TLS")
-    context.init(keyManagerFactory.getKeyManagers, null, new SecureRandom)
-
-    new HttpsConnectionContext(context)
-  }
-
-  private def withGrpcJavaServer()(block: => Unit): Assertion = {
-    try {
-      val server = new TestServiceServer
-      if (server.useTls)
-        println("\nUsing fake CA for TLS certificate. Test clients should expect host\n" +
-          "*.test.google.fr and our test CA. For the Java test client binary, use:\n" +
-          "--server_host_override=foo.test.google.fr --use_test_ca=true\n")
-
-      server.start()
-      try
-        block
-      finally
-        server.stop()
-      Succeeded
-    } catch {
       case NonFatal(t) => fail(t)
     }
   }
@@ -195,7 +103,7 @@ trait PendingCases {
 }
 
 trait ServerHandlerProvider extends PendingCases {
-  def serverHandlerFactory: Materializer => ExecutionContext => PartialFunction[HttpRequest, Future[HttpResponse]]
+  def server: GrpcServer[_]
 }
 
 trait  ClientTesterProvider extends PendingCases {
@@ -211,8 +119,9 @@ trait GrpcJavaPendingCases extends PendingCases {
     )
 }
 
-object GrpcJavaServer extends GrpcJavaPendingCases {
+object IoGrpcJavaServer extends GrpcJavaPendingCases with ServerHandlerProvider {
   val label: String = "grcp-java server"
+  val server = IoGrpcServer
 }
 
 object GrpcJavaClientTesterProvider extends ClientTesterProvider with GrpcJavaPendingCases {
@@ -223,7 +132,7 @@ object GrpcJavaClientTesterProvider extends ClientTesterProvider with GrpcJavaPe
 
 trait AkkaHttpServerProvider extends ServerHandlerProvider {
 
-  val label: String = "akka-grpc server"
+  val label: String = "akka-grpc server scala"
   val pendingCases =
     Set(
       "custom_metadata",
