+++ interop-tests/src/main/scala/com/lightbend/grpc/interop/GrpcServer.scala
@@ -1,0 +1,6 @@
+package com.lightbend.grpc.interop
+
+abstract class GrpcServer[T] {
+  def start(): T
+  def stop(binding: T): Unit
+}
