+++ interop-tests/src/test/scala/com/lightbend/grpc/interop/GrpcInteropSpec.scala
@@ -1,24 +1,21 @@
 package com.lightbend.grpc.interop
 
-import akka.http.scaladsl.model.{ HttpRequest, HttpResponse }
-import akka.stream.Materializer
 import io.grpc.testing.integration.test.{ AkkaGrpcClientTester, TestServiceServiceHandler }
 import io.grpc.testing.integration2.{ ClientTester, Settings }
 import org.scalatest._
 
-import scala.concurrent.{ ExecutionContext, Future }
+import scala.concurrent.ExecutionContext
 
 class GrpcInteropSpec extends WordSpec with GrpcInteropTests {
 
-  javaGrpcTests(GrpcJavaClientTesterProvider)
-  javaGrpcTests(AkkaHttpClientProvider)
+  grpcTests(IoGrpcJavaServer, GrpcJavaClientTesterProvider)
+  grpcTests(IoGrpcJavaServer, AkkaHttpClientProvider)
 
-  akkaGrpcTests(AkkaHttpServerProvider, GrpcJavaClientTesterProvider)
-  akkaGrpcTests(AkkaHttpServerProvider, AkkaHttpClientProvider)
+  grpcTests(AkkaHttpServerProvider, GrpcJavaClientTesterProvider)
+  grpcTests(AkkaHttpServerProvider, AkkaHttpClientProvider)
 
   object AkkaHttpServerProvider extends AkkaHttpServerProvider {
-    val serverHandlerFactory: Materializer => ExecutionContext => PartialFunction[HttpRequest, Future[HttpResponse]] =
-      implicit mat => implicit ec => TestServiceServiceHandler(new TestServiceImpl())
+    val server = AkkaGrpcServerScala(implicit mat => implicit ec => TestServiceServiceHandler(new TestServiceImpl()))
   }
 
   object AkkaHttpClientProvider extends AkkaClientTestProvider {
