+++ sbt-plugin/src/sbt-test/gen-scala-server/01-gen-basic-server/project/build.properties
@@ -1,0 +1,1 @@
+sbt.version=1.1.1
